pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-digest"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "praw>=7.7",
    "feedparser>=6.0",
    "httpx>=0.27",
    "jinja2>=3.1",
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
ai-digest = "ai_digest.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
